Move cache.py to logger.py
