+++ tfep/io/__init__.py
@@ -1,4 +1,4 @@
-from tfep.io.cache import TFEPCache
+from tfep.io.logger import TFEPCache
 from tfep.io.dataset import (
     DictDataset, MergedDataset, TrajectoryDataset,
     TrajectorySubset, get_subsampled_indices,
